requirements.txt
APScheduler==3.10.4
async-timeout==4.0.3
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
gunicorn==22.0.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
requests==2.32.2
six==1.16.0
tzlocal==5.2
urllib3==2.2.1
Werkzeug==3.0.3
zipp==3.19.0
